Removed all BeOS Dano headers, it is now self-hosting
