+++ ArpHeader/BeExp/HaikuBased/ColumnTypes.h
