+++ ArpHeader/BeExp/SampleCodeLicense/BitmapButton.h
