+++ ArpHeader/BeExp/SampleCodeLicense/BitmapTools.h
