+++ ArpHeader/BeExp/SampleCodeLicense/ColorTools.h
