+++ ArpHeader/BeExp/SampleCodeLicense/ResourceSet.h
