+++ ArpHeader/Sequitur/SeqApplication.h
@@ -6,7 +6,7 @@
 
 #include <app/Application.h>
 #include <app/Messenger.h>
-#include <experimental/DocApplication.h>
+#include <SelfWritten/DocApplication.h>
 #include <interface/Bitmap.h>
 #include <interface/Picture.h>
 #include <support/Locker.h>
