+++ ArpSrc/BeExp/HaikuBased/ColumnTypes.cpp
