+++ ArpSrc/BeExp/InSourceZip/AddOnManager.cpp
