+++ ArpSrc/BeExp/InSourceZip/Gehnaphore.cpp
