+++ ArpSrc/BeExp/InSourceZip/SearchPath.cpp
