+++ ArpSrc/BeExp/SampleCodeLicense/BitmapButton.cpp
