+++ ArpSrc/BeExp/SampleCodeLicense/BitmapTools.cpp
