+++ ArpSrc/BeExp/SampleCodeLicense/ColorTools.cpp
