+++ ArpSrc/BeExp/SampleCodeLicense/ResourceSet.cpp
