+++ ArpSrc/BeExp/SelfWritten/DocWindow.cpp
@@ -1,44 +1,47 @@
-#include <experimental/DocWindow.h>
-#include <stdio.h>
+#include <Entry.h>
 #include <Message.h>
-DocWindow::DocWindow(WindowRoster* r, entry_ref* ref, BRect frame,
+
+#include <stdio.h>
+
+#include <experimental/DocWindow.h>
+
+DocWindow::DocWindow(DocApplication* r, entry_ref* ref, BRect frame,
 	const char *title, window_look look, window_feel feel,
-	uint32 flags, uint32 workspace = B_CURRENT_WORKSPACE)
+	uint32 flags, uint32 workspace)
 	:
 	BWindow(frame, title, look, feel, flags, workspace)
 {
 	windowroster = r;
-	if(ref)
+	if (ref)
 		fileref = *ref;
-	else
-		fileref.name = "";
 	dirty = false;
 	printf("DocWindow::DocWindow(%s)\n", title);
 }
+
 
 DocWindow::~DocWindow()
 {
 	printf("DocWindow::~DocWindow()\n");
 }
 
+
 bool
 DocWindow::QuitRequested()
 {
-	printf("DocWindow::QuitRequested()\n");
-	return true;
+	return !dirty;
 }
 
 void
 DocWindow::MessageReceived(BMessage* msg)
 {
 	switch (msg->what) {
-		case 'Dsav':
-			if(strlen(fileref.name) > 0)
+		case DOC_WIN_SAVE:
+			if(untitled)
 				Save(new BEntry(&fileref));
 			else
 				SaveAs();
 			break;
-		case 'Dsas':
+		case DOC_WIN_SAVE_AS:
 			SaveAs();
 			break;
 		default:
@@ -47,11 +50,6 @@ DocWindow::MessageReceived(BMessage* msg)
 	}
 }
 
-void
-DocWindow::MenusBeginning()
-{
-	printf("DocWindow::MenusBeginning()\n");	
-}
 
 bool
 DocWindow::IsDirty()
@@ -59,11 +57,6 @@ DocWindow::IsDirty()
 	return dirty;
 }
 
-void
-DocWindow::EntryChanged(BMessage* msg)
-{
-	printf("DocWindow::EntryChanged()\n");
-}
 
 status_t
 DocWindow::Load(BEntry* e)
@@ -72,6 +65,7 @@ DocWindow::Load(BEntry* e)
 	return B_OK;
 }
 
+
 status_t
 DocWindow::Save(BEntry* e, const BMessage* args = 0)
 {
@@ -79,11 +73,14 @@ DocWindow::Save(BEntry* e, const BMessage* args = 0)
 	return B_ERROR;
 }
 
+
 void
 DocWindow::SaveAs()
 {
 	printf("DocWindow::SaveAs()\n");
+	untitled = false;
 }
+
 
 BFilePanel*
 DocWindow::CreateSavePanel() const
@@ -92,17 +89,13 @@ DocWindow::CreateSavePanel() const
 	return NULL;
 }
 
-void
-DocWindow::WindowFrame(BRect* proposed)
-{
-	*proposed = Frame();
-}
 
 void
-DocWindow::SetDirty(bool flag)
+DocWindow::SetDirty(bool setDirty)
 {
-	dirty = flag;
+	dirty = setDirty;
 }
+
 
 entry_ref
 DocWindow::FileRef() const
