+++ Sequitur/OldCode/SeqSongWindowWithAudioTest.cpp
@@ -6,7 +6,7 @@
 #include <RecentItems.h>
 #include <Roster.h>
 #include <experimental/BitmapButton.h>
-#include <experimental/DocApplication.h>
+#include <SelfWritten/DocApplication.h>
 #include <InterfaceKit.h>
 #include <kernel/fs_attr.h>
 #include <StorageKit.h>
